InfrastructureSystems

InfrastructureSystems

API documentation

InfrastructureSystems

Index

Exported

Internal

Index

InfrastructureSystems.DataFormatError

InfrastructureSystems.Deterministic

InfrastructureSystems.Deterministic

InfrastructureSystems.Deterministic

InfrastructureSystems.FlattenIteratorWrapper

InfrastructureSystems.Forecasts

InfrastructureSystems.Forecasts

InfrastructureSystems.InfrastructureSystemsInternal

InfrastructureSystems.InfrastructureSystemsInternal

InfrastructureSystems.LazyDictFromIterator

InfrastructureSystems.LogEvent

InfrastructureSystems.LogEventTracker

InfrastructureSystems.MultiLogger

InfrastructureSystems.MultiLogger

InfrastructureSystems.Probabilistic

InfrastructureSystems.Probabilistic

InfrastructureSystems.Probabilistic

InfrastructureSystems.ScenarioBased

InfrastructureSystems.ScenarioBased

InfrastructureSystems.ScenarioBased

InfrastructureSystems.TimeseriesFileMetadata

Base.close

Base.flush

Base.get

InfrastructureSystems._get_all_concrete_subtypes

InfrastructureSystems._validate_forecast

InfrastructureSystems.add_component!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecasts!

InfrastructureSystems.clear_forecasts!

InfrastructureSystems.compare_values

InfrastructureSystems.configure_logging

InfrastructureSystems.convert_type

InfrastructureSystems.convert_type

InfrastructureSystems.convert_type!

InfrastructureSystems.from_json

InfrastructureSystems.from_json

InfrastructureSystems.get_abstract_subtypes

InfrastructureSystems.get_all_concrete_subtypes

InfrastructureSystems.get_component

InfrastructureSystems.get_component

InfrastructureSystems.get_component

InfrastructureSystems.get_component

InfrastructureSystems.get_component_forecasts

InfrastructureSystems.get_component_types_raw

InfrastructureSystems.get_components

InfrastructureSystems.get_components_by_name

InfrastructureSystems.get_components_raw

InfrastructureSystems.get_concrete_subtypes

InfrastructureSystems.get_data

InfrastructureSystems.get_data

InfrastructureSystems.get_data

InfrastructureSystems.get_forecast_component_name

InfrastructureSystems.get_forecast_initial_times

InfrastructureSystems.get_forecast_value

InfrastructureSystems.get_forecast_value

InfrastructureSystems.get_forecasts

InfrastructureSystems.get_forecasts

InfrastructureSystems.get_forecasts

InfrastructureSystems.get_forecasts_horizon

InfrastructureSystems.get_forecasts_initial_time

InfrastructureSystems.get_forecasts_interval

InfrastructureSystems.get_forecasts_resolution

InfrastructureSystems.get_horizon

InfrastructureSystems.get_horizon

InfrastructureSystems.get_horizon

InfrastructureSystems.get_initial_time

InfrastructureSystems.get_initial_time

InfrastructureSystems.get_initial_time

InfrastructureSystems.get_internal

InfrastructureSystems.get_internal

InfrastructureSystems.get_internal

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_log_events

InfrastructureSystems.get_percentiles

InfrastructureSystems.get_period_columns

InfrastructureSystems.get_resolution

InfrastructureSystems.get_resolution

InfrastructureSystems.get_resolution

InfrastructureSystems.get_scenario_count

InfrastructureSystems.get_start_index

InfrastructureSystems.get_start_index

InfrastructureSystems.get_start_index

InfrastructureSystems.get_step_time

InfrastructureSystems.get_timeseries

InfrastructureSystems.get_timeseries_format

InfrastructureSystems.get_timestamp

InfrastructureSystems.get_unique_timestamps

InfrastructureSystems.get_uuid

InfrastructureSystems.get_value_columns

InfrastructureSystems.get_value_columns

InfrastructureSystems.increment_count

InfrastructureSystems.iterate_components

InfrastructureSystems.iterate_forecasts

InfrastructureSystems.make_forecasts

InfrastructureSystems.make_forecasts

InfrastructureSystems.make_forecasts

InfrastructureSystems.make_forecasts

InfrastructureSystems.open_file_logger

InfrastructureSystems.read_timeseries

InfrastructureSystems.read_timeseries

InfrastructureSystems.read_timeseries

InfrastructureSystems.read_timeseries

InfrastructureSystems.read_timeseries_metadata

InfrastructureSystems.remove_component!

InfrastructureSystems.remove_component!

InfrastructureSystems.remove_components!

InfrastructureSystems.remove_forecast!

InfrastructureSystems.replace_iterator

InfrastructureSystems.report_log_summary

InfrastructureSystems.report_log_summary

InfrastructureSystems.reset_iterator

InfrastructureSystems.separate_type_and_parameter_types

InfrastructureSystems.split_forecasts!

InfrastructureSystems.strip_module_names

InfrastructureSystems.supertypes

InfrastructureSystems.to_json

InfrastructureSystems.to_json

InfrastructureSystems.to_json

InfrastructureSystems.type_to_symbol

InfrastructureSystems.validate_components

InfrastructureSystems.validate_struct

JSON2.read

JSON2.write

Exported

InfrastructureSystems.DataFormatError — Type.

Thrown upon detection of user data that is not supported.

InfrastructureSystems.Deterministic — Type.

A deterministic forecast for a particular data field in a Component.

InfrastructureSystems.Deterministic — Method.

Constructs Deterministic after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.Deterministic — Method.

Constructs Deterministic from a InfrastructureSystemsType, label, and TimeArray.

InfrastructureSystems.FlattenIteratorWrapper — Type.

Wrapper around Iterators.Flatten to provide total length.

InfrastructureSystems.Forecasts — Type.

Container for forecasts and their metadata..

InfrastructureSystems.Forecasts — Method.

Partially constructs Forecasts from JSON. Forecasts are not constructed.

InfrastructureSystems.InfrastructureSystemsInternal — Type.

Internal storage common to InfrastructureSystems types.

InfrastructureSystems.InfrastructureSystemsInternal — Method.

Creates PowerSystemInternal with a UUID.

InfrastructureSystems.LogEvent — Type.

Contains information describing a log event.

InfrastructureSystems.LogEventTracker — Type.

LogEventTracker(Tuple{Logging.LogLevel})

Tracks counts of all log events by level.

Examples

LogEventTracker()
LogEventTracker((Logging.Info, Logging.Warn, Logging.Error))

InfrastructureSystems.MultiLogger — Type.

MultiLogger(Array{AbstractLogger}, Union{LogEventTracker, Nothing})

Redirects log events to multiple loggers. The primary use case is to allow logging to both a file and the console. Secondarily, it can track the counts of all log messages.

Example

MultiLogger([ConsoleLogger(stderr), SimpleLogger(stream)], LogEventTracker())

InfrastructureSystems.MultiLogger — Method.

Creates a MultiLogger with no event tracking.

Example

MultiLogger([ConsoleLogger(stderr), SimpleLogger(stream)])

InfrastructureSystems.Probabilistic — Type.

A Probabilistic forecast for a particular data field in a Component.

InfrastructureSystems.Probabilistic — Method.

Constructs Probabilistic Forecast after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.Probabilistic — Method.

Constructs Probabilistic after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.ScenarioBased — Type.

A Discrete Scenario Based forecast for a particular data field in a Component.

InfrastructureSystems.ScenarioBased — Method.

Constructs ScenarioBased Forecast after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.ScenarioBased — Method.

Constructs ScenarioBased Forecast after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.TimeseriesFileMetadata — Type.

Describes how to construct forecasts from raw timeseries data files.

InfrastructureSystems.add_component! — Method.

add_component!(
               components::Components,
               component::T;
               kwargs...
              ) where T <: InfrastructureSystemsType

Add a component.

Throws ArgumentError if the component's name is already stored for its concrete type.

Throws InvalidRange if any of the component's field values are outside of defined valid range.

InfrastructureSystems.add_forecast! — Function.

add_forecast!(
              data::SystemData,
              filename::AbstractString,
              component::InfrastructureSystemsType,
              label::AbstractString,
              scaling_factor::Union{String, Float64}=1.0,
             )

Add a forecast from a CSV file.

See TimeseriesFileMetadata for description of scaling_factor.

InfrastructureSystems.add_forecast! — Function.

add_forecast!(
              data::SystemData,
              ta::TimeSeries.TimeArray,
              component::InfrastructureSystemsType,
              label::AbstractString,
              scaling_factor::Union{String, Float64}=1.0,
             )

Add a forecast to a system from a TimeSeries.TimeArray.

See TimeseriesFileMetadata for description of scaling_factor.

InfrastructureSystems.add_forecast! — Function.

add_forecast!(
              data::SystemData,
              df::DataFrames.DataFrame,
              component::InfrastructureSystemsType,
              label::AbstractString,
              scaling_factor::Union{String, Float64}=1.0;
              timestamp=:timestamp,
             )

Add a forecast to a system from a DataFrames.DataFrame.

See TimeseriesFileMetadata for description of scaling_factor.

InfrastructureSystems.add_forecast! — Method.

add_forecast!(data::SystemData, forecasts)

Add forecasts.

Arguments

data::SystemData: infrastructure

forecast: Any object of subtype forecast

Throws ArgumentError if the forecast's component is not stored in the system.

InfrastructureSystems.add_forecasts! — Method.

add_forecasts!(data::SystemData, forecasts)

Add forecasts.

Arguments

data::SystemData: system

forecasts: iterable (array, iterator, etc.) of Forecast values

Throws DataFormatError if

A component-label pair is not unique within a forecast array.

A forecast has a different resolution than others.

A forecast has a different horizon than others.

Throws ArgumentError if the forecast's component is not stored in the system.

InfrastructureSystems.clear_forecasts! — Method.

clear_forecasts!(forecasts::Forecasts)

Remove all forecast objects from a Forecasts

InfrastructureSystems.configure_logging — Method.

configure_logging([console, console_stream, console_level,
                   file, filename, file_level, file_mode,
                   tracker, set_global])

Creates console and file loggers per caller specification and returns a MultiLogger.

Note: If logging to a file users must call Base.close() on the returned MultiLogger to ensure that all events get flushed.

Arguments

console::Bool=true: create console logger

console_stream::IOStream=stderr: stream for console logger

console_level::Logging.LogLevel=Logging.Error: level for console messages

file::Bool=true: create file logger

filename::String=log.txt: log file

file_level::Logging.LogLevel=Logging.Info: level for file messages

file_mode::String=w+: mode used when opening log file

tracker::Union{LogEventTracker, Nothing}=LogEventTracker(): optionally track log events

set_global::Bool=true: set the created logger as the global logger

Example

logger = configure_logging(filename="mylog.txt")

InfrastructureSystems.get_component — Method.

Get Deterministic component.

InfrastructureSystems.get_component — Method.

Get Probabilistic component.

InfrastructureSystems.get_component — Method.

Get ScenarioBased component.

InfrastructureSystems.get_component — Method.

get_component(
              ::Type{T},
              components::Components,
              name::AbstractString
             )::Union{T, Nothing} where T <: InfrastructureSystemsType

Get the component of concrete type T with name. Returns nothing if no component matches.

See get_components_by_name if the concrete type is unknown.

Throws ArgumentError if T is not a concrete type.

InfrastructureSystems.get_components — Method.

get_components(
               ::Type{T},
               components::Components,
              )::FlattenIteratorWrapper{T} where T <: InfrastructureSystemsType

Returns an iterator of components. T can be concrete or abstract. Call collect on the result if an array is desired.

See also: iterate_components

InfrastructureSystems.get_components_by_name — Method.

get_components_by_name(
                       ::Type{T},
                       components::Components,
                       name::AbstractString
                      )::Vector{T} where T <: InfrastructureSystemsType

Get the components of abstract type T with name. Note that InfrastructureSystems enforces unique names on each concrete type but not across concrete types.

See get_component if the concrete type is known.

Throws ArgumentError if T is not an abstract type.

InfrastructureSystems.get_data — Method.

Get Deterministic data.

InfrastructureSystems.get_data — Method.

Get Probabilistic data.

InfrastructureSystems.get_data — Method.

Get ScenarioBased data.

InfrastructureSystems.get_forecast_component_name — Method.

Get Forecast component name.

InfrastructureSystems.get_forecast_initial_times — Method.

get_forecast_initial_times(forecasts::Forecasts)::Vector{Dates.DateTime}

Return sorted forecast initial times.

InfrastructureSystems.get_forecast_value — Method.

gets the value of a Deterministic forecast at a given index or DateTime timestamp

InfrastructureSystems.get_forecast_value — Method.

gets the array of values of a forecast at a given index or DateTime timestamp

InfrastructureSystems.get_forecasts — Function.

get_forecasts(
              forecasts::Forecasts,
              initial_time::Dates.DateTime,
              components_iterator,
              label::Union{String, Nothing}=nothing,
             )

Arguments

forecasts::Forecasts: system

initial_time::Dates.DateTime: time designator for the forecast

components_iter: iterable (array, iterator, etc.) of Component values

label::Union{String, Nothing}: forecast label or nothing

Return forecasts of any type <: Forecast that match the components and label.

This method is slower than the first version because it has to compare components and label as well as build a new vector.

Throws ArgumentError if eltype(components_iterator) is a concrete type and no forecast is found for a component.

InfrastructureSystems.get_forecasts — Method.

get_forecasts(
              ::Type{T},
              forecasts::Forecasts,
              initial_time::Dates.DateTime,
              components_iterator,
              label::Union{String, Nothing}=nothing,
             )::Vector{Forecast}

Arguments

forecasts::Forecasts: system

initial_time::Dates.DateTime: time designator for the forecast

components_iter: iterable (array, iterator, etc.) of Component values

label::Union{String, Nothing}: forecast label or nothing

Return forecasts that match the components and label.

This method is slower than the first version because it has to compare components and label as well as build a new vector.

Throws ArgumentError if eltype(components_iterator) is a concrete type and no forecast is found for a component.

InfrastructureSystems.get_forecasts — Method.

get_forecasts(::Type{T}, forecasts::Forecasts, initial_time::Dates.DateTime)

Return an iterator of forecasts. T can be concrete or abstract.

Call collect on the result if an array is desired.

This method is fast and efficient because it returns an iterator to existing vectors.

Examples

iter = PowerSystems.get_forecasts(Deterministic{RenewableFix}, forecasts, initial_time)
iter = PowerSystems.get_forecasts(Forecast, forecasts, initial_time)
forecasts = PowerSystems.get_forecasts(Forecast, forecasts, initial_time) |> collect
forecasts = collect(PowerSystems.get_forecasts(Forecast, forecasts))

See also: iterate_forecasts

InfrastructureSystems.get_forecasts_horizon — Method.

Return the horizon for all forecasts.

InfrastructureSystems.get_forecasts_initial_time — Method.

Return the earliest initial_time for a forecast.

InfrastructureSystems.get_forecasts_interval — Method.

Return the interval for all forecasts.

InfrastructureSystems.get_forecasts_resolution — Method.

Return the resolution for all forecasts.

InfrastructureSystems.get_horizon — Method.

Get Deterministic horizon.

InfrastructureSystems.get_horizon — Method.

Get Probabilistic horizon.

InfrastructureSystems.get_horizon — Method.

Get ScenarioBased horizon.

InfrastructureSystems.get_initial_time — Method.

Get Deterministic initial_time.

InfrastructureSystems.get_initial_time — Method.

Get Probabilistic initial_time.

InfrastructureSystems.get_initial_time — Method.

Get ScenarioBased initial_time.

InfrastructureSystems.get_internal — Method.

Get Deterministic internal.

InfrastructureSystems.get_internal — Method.

Get Probabilistic internal.

InfrastructureSystems.get_internal — Method.

Get ScenarioBased internal.

InfrastructureSystems.get_label — Method.

Get Deterministic label.

InfrastructureSystems.get_label — Method.

Get Probabilistic label.

InfrastructureSystems.get_label — Method.

Get ScenarioBased label.

InfrastructureSystems.get_log_events — Method.

Returns an iterable of log events for a level.

InfrastructureSystems.get_percentiles — Method.

Get Probabilistic percentiles.

InfrastructureSystems.get_resolution — Method.

Get Deterministic resolution.

InfrastructureSystems.get_resolution — Method.

Get Probabilistic resolution.

InfrastructureSystems.get_resolution — Method.

Get ScenarioBased resolution.

InfrastructureSystems.get_scenario_count — Method.

Get ScenarioBased scenario_count.

InfrastructureSystems.get_start_index — Method.

Get Deterministic start_index.

InfrastructureSystems.get_start_index — Method.

Get Probabilistic start_index.

InfrastructureSystems.get_start_index — Method.

Get ScenarioBased start_index.

InfrastructureSystems.get_timeseries — Method.

get_timeseries(forecast::Forecast)

Return the timeseries for the forecast.

Note: timeseries data is stored in TimeSeries.TimeArray objects. TimeArray does not currently support Base.view, so calling this function results in a memory allocation and copy. Tracked in https://github.com/JuliaStats/TimeSeries.jl/issues/419.

InfrastructureSystems.iterate_components — Method.

iterate_components(obj) where T <: InfrastructureSystemsType

Iterates over all components.

Examples

for component in iterate_components(obj)
    @show component
end

See also: get_components

InfrastructureSystems.iterate_forecasts — Method.

Iterates over all forecasts in order of initial time.

Examples

for forecast in iterate_forecasts(forecasts)
    @show forecast
end

See also: get_forecasts

InfrastructureSystems.make_forecasts — Method.

make_forecasts(forecast::Deterministic, interval::Dates.Period, horizon::Int)

Make a vector of forecasts by incrementing through a forecast by interval and horizon.

InfrastructureSystems.make_forecasts — Method.

make_forecasts(data::SystemData, metadata_file::AbstractString; resolution=nothing)

Makes forecasts from a metadata file.

Arguments

data::SystemData: system

metadata_file::AbstractString: path to metadata file

mod::Module: calling module

resolution::{Nothing, Dates.Period}: skip any forecasts that don't match this resolution

See TimeseriesFileMetadata for description of what the file should contain.

InfrastructureSystems.make_forecasts — Method.

make_forecasts(data::SystemData, timeseries_metadata::Vector{TimeseriesFileMetadata};
               resolution=nothing)

Return a vector of forecasts from a vector of TimeseriesFileMetadata values.

Arguments

data::SystemData: system

timeseries_metadata::Vector{TimeseriesFileMetadata}: metadata values

resolution::{Nothing, Dates.Period}: skip any forecasts that don't match this resolution

InfrastructureSystems.make_forecasts — Method.

make_forecasts(forecast::FlattenIteratorWrapper{T},
                interval::Dates.Period, horizon::Int) where T <: Forecast

Make an iterator of forecasts by incrementing through a vector of forecasts by interval and horizon.

InfrastructureSystems.open_file_logger — Function.

open_file_logger(func, filename[, level, mode])

Opens a file logger using Logging.SimpleLogger.

Example

open_file_logger("log.txt", Logging.Info) do logger
    global_logger(logger)
    @info "hello world"
end

InfrastructureSystems.read_timeseries_metadata — Method.

Reads forecast metadata and fixes relative paths to the data files.

InfrastructureSystems.remove_component! — Method.

remove_component!(
                  components::Components,
                  component::T,
                 ) where T <: InfrastructureSystemsType

Remove a component by its value.

Throws ArgumentError if the component is not stored.

InfrastructureSystems.remove_component! — Method.

remove_component!(
                  ::Type{T},
                  components::Components,
                  name::AbstractString,
                  ) where T <: InfrastructureSystemsType

Remove a component by its name.

Throws ArgumentError if the component is not stored.

InfrastructureSystems.remove_components! — Method.

remove_components!(
                   ::Type{T},
                   components::Components,
                  ) where T <: InfrastructureSystemsType

Remove all components of type T.

Throws ArgumentError if the type is not stored.

InfrastructureSystems.remove_forecast! — Method.

remove_forecast(forecasts::Forecasts, forecast::Forecast)

Remove the forecast from the system.

Throws ArgumentError if the forecast is not stored.

InfrastructureSystems.report_log_summary — Method.

Returns a summary of log event counts by level.

InfrastructureSystems.report_log_summary — Method.

Returns a summary of log event counts by level.

InfrastructureSystems.split_forecasts! — Method.

split_forecasts!(data::SystemData,
                 forecasts,
                 interval::Dates.Period,
                 horizon::Int) where T <: Forecast

Replaces system forecasts with a set of forecasts by incrementing through an iterable set of forecasts by interval and horizon.

InfrastructureSystems.validate_struct — Method.

Validates a struct.

Internal

InfrastructureSystems.LazyDictFromIterator — Method.

LazyDictFromIterator creates a dictionary from an iterator, but only increments the iterator and adds items to the dictionary as it needs them. In the worst case it is identical to creating a dictionary by iterating over the entire list. Each V should have a K member.

Arguments

K: type of the dictionary keys

V: type of the dictionary values

iter: any object implementing the Iterator interface

getter::Function: method to call on V to get its K

Base.close — Method.

Ensures that any file streams are flushed and closed.

Base.flush — Method.

Flush any file streams.

Base.get — Method.

Returns the item mapped to key. If the key is already stored then it will be returned with a dictionary lookup. If it has not been stored then iterate over the list until it is found.

Returns nothing if key is not found.

InfrastructureSystems._get_all_concrete_subtypes — Method.

Recursively builds a vector of subtypes.

InfrastructureSystems._validate_forecast — Method.

Checks that the component exists in data and the UUID's match.

InfrastructureSystems.compare_values — Method.

Recursively compares immutable struct values by performing == on each field in the struct. When performing == on values of immutable structs Julia will perform === on each field. This will return false if any field is mutable even if the contents are the same. So, comparison of any InfrastructureSystems type with an array will fail.

This is an unresolved Julia issue. Refer to https://github.com/JuliaLang/julia/issues/4648.

An option is to overload == for all subtypes of PowerSystemType. That may not be appropriate in all cases. Until the Julia developers decide on a solution, this function is provided for convenience for specific comparisons.

InfrastructureSystems.convert_type! — Method.

Converts forecast JSON data to Forecasts.

InfrastructureSystems.convert_type — Method.

Converts an object deserialized from JSON into a Julia type, such as NamedTuple, to an instance of T. Similar to Base.convert, but not a viable replacement.

InfrastructureSystems.convert_type — Method.

Creates a Forecast object by decoding the data that was in JSON. This data stores the values for the field contributingdevices as UUIDs, so this will lookup each device in devices.

InfrastructureSystems.from_json — Method.

Deserializes a InfrastructureSystemsType from a JSON filename.

InfrastructureSystems.from_json — Method.

Deserializes a InfrastructureSystemsType from String or IO.

InfrastructureSystems.get_abstract_subtypes — Method.

Returns an array of abstract types that are direct subtypes of T.

InfrastructureSystems.get_all_concrete_subtypes — Method.

Returns an array of all concrete subtypes of T.

InfrastructureSystems.get_component_forecasts — Method.

get_component_forecasts(
                        ::Type{T},
                        forecasts::Forecasts,
                        initial_time::Dates.DateTime,
                       ) where T <: InfrastructureSystemsType

Get the forecasts of a component of type T with initial_time. The resulting container can contain Forecasts of dissimilar types.

Throws ArgumentError if T is not a concrete type.

See also: get_component

InfrastructureSystems.get_component_types_raw — Method.

Return an iterable of component types deserialized from JSON.

InfrastructureSystems.get_components_raw — Method.

Return an iterable of components as NamedTuples deserialized from JSON.

InfrastructureSystems.get_concrete_subtypes — Method.

Returns an array of concrete types that are direct subtypes of T.

InfrastructureSystems.get_period_columns — Method.

Return the column names that specify the Period.

InfrastructureSystems.get_step_time — Method.

Return a DateTime for the step between values as specified by the period in the file.

InfrastructureSystems.get_timeseries_format — Method.

Return the timeseries format used in the CSV file.

InfrastructureSystems.get_timestamp — Method.

Return a Dates.DateTime for the row in the CSV file.

InfrastructureSystems.get_unique_timestamps — Method.

Return a vector of dicts of unique timestamps and their counts.

InfrastructureSystems.get_uuid — Method.

Gets the UUID for any PowerSystemType.

InfrastructureSystems.get_value_columns — Method.

Return the column names with values.

InfrastructureSystems.get_value_columns — Method.

Return the column names with values (components).

InfrastructureSystems.increment_count — Method.

Increments the count of a log event.

InfrastructureSystems.read_timeseries — Function.

read_timeseries(file_path::AbstractString, component_name=nothing)

Return a TimeArray from a CSV file.

Pass component_name when the file does not have the component name in a column header.

InfrastructureSystems.read_timeseries — Method.

This version of the function supports the format where there is no column header for a component, so the component_name must be passed in.

InfrastructureSystems.read_timeseries — Method.

This version of the function only has component_name to match the interface. It is unused.

Set start_datetime as a keyword argument for the starting timestamp, otherwise the current day is used.

InfrastructureSystems.read_timeseries — Method.

Return a TimeSeries.TimeArray representing the CSV file.

This version of the function only has component_name to match the interface. It is unused.

InfrastructureSystems.replace_iterator — Method.

Replace the iterator, maintaining the cached dict.

InfrastructureSystems.reset_iterator — Method.

Reset the iterator for cases where underlying arrays have changed.

InfrastructureSystems.separate_type_and_parameter_types — Method.

Return a Tuple of type and parameter types for cases where a parametric type has been encoded as a string. If the type is not parameterized then just return the type.

InfrastructureSystems.strip_module_names — Method.

Strips the module name(s) off of a type.

InfrastructureSystems.supertypes — Method.

Returns an array of all super types of T.

InfrastructureSystems.to_json — Method.

JSON Serializes a InfrastructureSystemsType to an IO stream in JSON.

InfrastructureSystems.to_json — Method.

Serializes a InfrastructureSystemsType to a JSON file.

InfrastructureSystems.to_json — Method.

Serializes a InfrastructureSystemsType to a JSON string.

InfrastructureSystems.type_to_symbol — Method.

Converts a DataType to a Symbol, stripping off the module name(s).

InfrastructureSystems.validate_components — Method.

validate_components(components::Components)

Iterates over all components and throws InvalidRange if any of the component's field values are outside of defined valid range.

JSON2.read — Method.

Enables JSON deserialization of TimeSeries.TimeArray. The default implementation fails because the data field is defined as an AbstractArray. Deserialization can't determine the actual concrete type.

JSON2.write — Method.

Enables JSON deserialization of Dates.Period. The default implementation fails because the field is defined as abstract. Encode the type when serializing so that the correct value can be deserialized.