InfrastructureSystems

InfrastructureSystems

API documentation

InfrastructureSystems

Index

Exported

Internal

Index

InfrastructureSystems.DataFormatError

InfrastructureSystems.Deterministic

InfrastructureSystems.Deterministic

InfrastructureSystems.DeterministicInternal

InfrastructureSystems.FlattenIteratorWrapper

InfrastructureSystems.Forecast

InfrastructureSystems.ForecastInternal

InfrastructureSystems.ForecastMetadata

InfrastructureSystems.Forecasts

InfrastructureSystems.Hdf5TimeSeriesStorage

InfrastructureSystems.Hdf5TimeSeriesStorage

InfrastructureSystems.InMemoryTimeSeriesStorage

InfrastructureSystems.InfrastructureSystemsInternal

InfrastructureSystems.InfrastructureSystemsInternal

InfrastructureSystems.LazyDictFromIterator

InfrastructureSystems.LogEvent

InfrastructureSystems.LogEventTracker

InfrastructureSystems.MultiLogger

InfrastructureSystems.MultiLogger

InfrastructureSystems.Probabilistic

InfrastructureSystems.Probabilistic

InfrastructureSystems.ProbabilisticInternal

InfrastructureSystems.ScenarioBased

InfrastructureSystems.ScenarioBased

InfrastructureSystems.ScenarioBasedInternal

InfrastructureSystems.SystemData

InfrastructureSystems.TimeSeriesStorage

InfrastructureSystems.TimeseriesFileMetadata

Base.close

Base.flush

Base.get

InfrastructureSystems._get_all_concrete_subtypes

InfrastructureSystems._remove_item!

InfrastructureSystems._split_forecast

InfrastructureSystems._validate_component

InfrastructureSystems.add_component!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecast!

InfrastructureSystems.add_forecasts!

InfrastructureSystems.add_forecasts!

InfrastructureSystems.check_contiguous_forecasts

InfrastructureSystems.check_forecast_consistency

InfrastructureSystems.clear_components!

InfrastructureSystems.compare_values

InfrastructureSystems.configure_logging

InfrastructureSystems.convert_type

InfrastructureSystems.convert_type

InfrastructureSystems.deserialize_components

InfrastructureSystems.from

InfrastructureSystems.from_file

InfrastructureSystems.from_json

InfrastructureSystems.from_json

InfrastructureSystems.generate_initial_times

InfrastructureSystems.generate_initial_times

InfrastructureSystems.get_abstract_subtypes

InfrastructureSystems.get_all_concrete_subtypes

InfrastructureSystems.get_component

InfrastructureSystems.get_component_types_raw

InfrastructureSystems.get_components

InfrastructureSystems.get_components_by_name

InfrastructureSystems.get_components_raw

InfrastructureSystems.get_concrete_subtypes

InfrastructureSystems.get_data

InfrastructureSystems.get_data

InfrastructureSystems.get_data

InfrastructureSystems.get_forecast

InfrastructureSystems.get_forecast

InfrastructureSystems.get_forecast_counts

InfrastructureSystems.get_forecast_values

InfrastructureSystems.get_forecasts_horizon

InfrastructureSystems.get_forecasts_interval

InfrastructureSystems.get_forecasts_resolution

InfrastructureSystems.get_horizon

InfrastructureSystems.get_horizon

InfrastructureSystems.get_horizon

InfrastructureSystems.get_initial_time

InfrastructureSystems.get_initial_time

InfrastructureSystems.get_initial_time

InfrastructureSystems.get_internal

InfrastructureSystems.get_internal

InfrastructureSystems.get_internal

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_label

InfrastructureSystems.get_log_events

InfrastructureSystems.get_percentiles

InfrastructureSystems.get_percentiles

InfrastructureSystems.get_period_columns

InfrastructureSystems.get_resolution

InfrastructureSystems.get_resolution

InfrastructureSystems.get_resolution

InfrastructureSystems.get_scenario_count

InfrastructureSystems.get_scenario_count

InfrastructureSystems.get_step_time

InfrastructureSystems.get_time_series_uuid

InfrastructureSystems.get_time_series_uuid

InfrastructureSystems.get_time_series_uuid

InfrastructureSystems.get_timeseries_format

InfrastructureSystems.get_timestamp

InfrastructureSystems.get_unique_timestamps

InfrastructureSystems.get_uuid

InfrastructureSystems.get_value_columns

InfrastructureSystems.get_value_columns

InfrastructureSystems.head

InfrastructureSystems.increment_count

InfrastructureSystems.iterate_components

InfrastructureSystems.make_forecast!

InfrastructureSystems.open_file_logger

InfrastructureSystems.prepare_for_removal!

InfrastructureSystems.prepare_for_serialization!

InfrastructureSystems.read_time_series

InfrastructureSystems.read_time_series

InfrastructureSystems.read_time_series

InfrastructureSystems.read_time_series

InfrastructureSystems.read_time_series_metadata

InfrastructureSystems.remove_component!

InfrastructureSystems.remove_component!

InfrastructureSystems.remove_components!

InfrastructureSystems.remove_forecast!

InfrastructureSystems.remove_forecast_internal!

InfrastructureSystems.replace_iterator

InfrastructureSystems.report_log_summary

InfrastructureSystems.report_log_summary

InfrastructureSystems.reset_iterator

InfrastructureSystems.separate_type_and_parameter_types

InfrastructureSystems.set_component!

InfrastructureSystems.strip_module_name

InfrastructureSystems.supertypes

InfrastructureSystems.tail

InfrastructureSystems.to

InfrastructureSystems.to_json

InfrastructureSystems.to_json

InfrastructureSystems.to_json

InfrastructureSystems.type_to_symbol

InfrastructureSystems.validate_components

InfrastructureSystems.validate_struct

InfrastructureSystems.when

JSON2.read

JSON2.write

Exported

InfrastructureSystems.LogEvent — Type.

Contains information describing a log event.

InfrastructureSystems.LogEventTracker — Type.

LogEventTracker(Tuple{Logging.LogLevel})

Tracks counts of all log events by level.

Examples

LogEventTracker()
LogEventTracker((Logging.Info, Logging.Warn, Logging.Error))

InfrastructureSystems.MultiLogger — Type.

MultiLogger(Array{AbstractLogger}, Union{LogEventTracker, Nothing})

Redirects log events to multiple loggers. The primary use case is to allow logging to both a file and the console. Secondarily, it can track the counts of all log messages.

Example

MultiLogger([ConsoleLogger(stderr), SimpleLogger(stream)], LogEventTracker())

InfrastructureSystems.MultiLogger — Method.

Creates a MultiLogger with no event tracking.

Example

MultiLogger([ConsoleLogger(stderr), SimpleLogger(stream)])

InfrastructureSystems.configure_logging — Method.

configure_logging([console, console_stream, console_level,
                   file, filename, file_level, file_mode,
                   tracker, set_global])

Creates console and file loggers per caller specification and returns a MultiLogger.

Note: If logging to a file users must call Base.close() on the returned MultiLogger to ensure that all events get flushed.

Arguments

console::Bool=true: create console logger

console_stream::IOStream=stderr: stream for console logger

console_level::Logging.LogLevel=Logging.Error: level for console messages

file::Bool=true: create file logger

filename::String=log.txt: log file

file_level::Logging.LogLevel=Logging.Info: level for file messages

file_mode::String=w+: mode used when opening log file

tracker::Union{LogEventTracker, Nothing}=LogEventTracker(): optionally track log events

set_global::Bool=true: set the created logger as the global logger

Example

logger = configure_logging(filename="mylog.txt")

InfrastructureSystems.get_data — Method.

Get Deterministic data.

InfrastructureSystems.get_data — Method.

Get Probabilistic data.

InfrastructureSystems.get_data — Method.

Get ScenarioBased data.

InfrastructureSystems.get_horizon — Method.

Get DeterministicInternal horizon.

InfrastructureSystems.get_horizon — Method.

Get ProbabilisticInternal horizon.

InfrastructureSystems.get_horizon — Method.

Get ScenarioBasedInternal horizon.

InfrastructureSystems.get_initial_time — Method.

Get DeterministicInternal initial_time.

InfrastructureSystems.get_initial_time — Method.

Get ProbabilisticInternal initial_time.

InfrastructureSystems.get_initial_time — Method.

Get ScenarioBasedInternal initial_time.

InfrastructureSystems.get_internal — Method.

Get DeterministicInternal internal.

InfrastructureSystems.get_internal — Method.

Get ProbabilisticInternal internal.

InfrastructureSystems.get_internal — Method.

Get ScenarioBasedInternal internal.

InfrastructureSystems.get_label — Method.

Get DeterministicInternal label.

InfrastructureSystems.get_label — Method.

Get Deterministic label.

InfrastructureSystems.get_label — Method.

Get ProbabilisticInternal label.

InfrastructureSystems.get_label — Method.

Get Probabilistic label.

InfrastructureSystems.get_label — Method.

Get ScenarioBasedInternal label.

InfrastructureSystems.get_label — Method.

Get ScenarioBased label.

InfrastructureSystems.get_log_events — Method.

Returns an iterable of log events for a level.

InfrastructureSystems.get_percentiles — Method.

Get ProbabilisticInternal percentiles.

InfrastructureSystems.get_percentiles — Method.

Get Probabilistic percentiles.

InfrastructureSystems.get_resolution — Method.

Get DeterministicInternal resolution.

InfrastructureSystems.get_resolution — Method.

Get ProbabilisticInternal resolution.

InfrastructureSystems.get_resolution — Method.

Get ScenarioBasedInternal resolution.

InfrastructureSystems.get_scenario_count — Method.

Get ScenarioBasedInternal scenario_count.

InfrastructureSystems.get_scenario_count — Method.

Get ScenarioBased scenario_count.

InfrastructureSystems.get_time_series_uuid — Method.

Get DeterministicInternal timeseriesuuid.

InfrastructureSystems.get_time_series_uuid — Method.

Get ProbabilisticInternal timeseriesuuid.

InfrastructureSystems.get_time_series_uuid — Method.

Get ScenarioBasedInternal timeseriesuuid.

InfrastructureSystems.open_file_logger — Function.

open_file_logger(func, filename[, level, mode])

Opens a file logger using Logging.SimpleLogger.

Example

open_file_logger("log.txt", Logging.Info) do logger
    global_logger(logger)
    @info "hello world"
end

InfrastructureSystems.report_log_summary — Method.

Returns a summary of log event counts by level.

InfrastructureSystems.report_log_summary — Method.

Returns a summary of log event counts by level.

Internal

InfrastructureSystems.DataFormatError — Type.

Thrown upon detection of user data that is not supported.

InfrastructureSystems.Deterministic — Type.

A deterministic forecast for a particular data field in a Component.

InfrastructureSystems.Deterministic — Method.

Deterministic(label::String,
              resolution::Dates.Period,
              initial_time::Dates.DateTime,
              time_steps::Int)

Constructs Deterministic after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.DeterministicInternal — Type.

A deterministic forecast for a particular data field in a Component.

InfrastructureSystems.FlattenIteratorWrapper — Type.

Wrapper around Iterators.Flatten to provide total length.

InfrastructureSystems.Forecast — Type.

Abstract type for forecasts supplied to users. They are not stored in a system. Instead, they are generated on demand for the user. Users can create them. The system will convert them to a subtype of ForecastInternal for storage. Time series data is stored as a field, so reads will always be from memory.

InfrastructureSystems.ForecastInternal — Type.

Abstract type for forecasts that are stored in a system. Users never create them or get access to them. Stores references to time series data, so a disk read may be required for access.

InfrastructureSystems.ForecastMetadata — Type.

Container for forecasts and their metadata..

InfrastructureSystems.Forecasts — Type.

Forecast container for a component.

InfrastructureSystems.Hdf5TimeSeriesStorage — Type.

Stores all time series data in an HDF5 file.

The file used is assumed to be temporary and will be automatically deleted when there are no more references to the storage object.

InfrastructureSystems.Hdf5TimeSeriesStorage — Method.

Constructs Hdf5TimeSeriesStorage by creating a new file.

InfrastructureSystems.InMemoryTimeSeriesStorage — Type.

Stores all time series data in memory.

InfrastructureSystems.InfrastructureSystemsInternal — Type.

Internal storage common to InfrastructureSystems types.

InfrastructureSystems.InfrastructureSystemsInternal — Method.

Creates PowerSystemInternal with a UUID.

InfrastructureSystems.LazyDictFromIterator — Method.

LazyDictFromIterator creates a dictionary from an iterator, but only increments the iterator and adds items to the dictionary as it needs them. In the worst case it is identical to creating a dictionary by iterating over the entire list. Each V should have a K member.

Arguments

K: type of the dictionary keys

V: type of the dictionary values

iter: any object implementing the Iterator interface

getter::Function: method to call on V to get its K

InfrastructureSystems.Probabilistic — Type.

A Probabilistic forecast for a particular data field in a Component.

InfrastructureSystems.Probabilistic — Method.

Probabilistic(
              label::String,
              resolution::Dates.Period,
              initial_time::Dates.DateTime,
              percentiles::Vector{Float64},
              time_steps::Int,
             )

Constructs Probabilistic after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.ProbabilisticInternal — Type.

A Probabilistic forecast for a particular data field in a Component.

InfrastructureSystems.ScenarioBased — Type.

A Discrete Scenario Based forecast for a particular data field in a Component.

InfrastructureSystems.ScenarioBased — Method.

Constructs ScenarioBased Forecast after constructing a TimeArray from initialtime and timesteps.

InfrastructureSystems.ScenarioBasedInternal — Type.

A Discrete Scenario Based forecast for a particular data field in a Component.

InfrastructureSystems.SystemData — Method.

SystemData(filename::AbstractString)

Deserialize SystemData from a JSON file.

InfrastructureSystems.TimeSeriesStorage — Type.

Abstract type for time series storage implementations.

All subtypes must implement:

addtimeseries!

removetimeseries!

gettimeseries

cleartimeseries!

getnumtime_series

InfrastructureSystems.TimeseriesFileMetadata — Type.

Describes how to construct forecasts from raw timeseries data files.

Base.close — Method.

Ensures that any file streams are flushed and closed.

Base.flush — Method.

Flush any file streams.

Base.get — Method.

Returns the item mapped to key. If the key is already stored then it will be returned with a dictionary lookup. If it has not been stored then iterate over the list until it is found.

Returns nothing if key is not found.

InfrastructureSystems._get_all_concrete_subtypes — Method.

Recursively builds a vector of subtypes.

InfrastructureSystems._remove_item! — Method.

Removes value from the dataset called name. Returns true if the array is empty afterwards.

InfrastructureSystems._split_forecast — Method.

Creates a new forecast from an existing forecast with a split TimeArray.

InfrastructureSystems._validate_component — Method.

Checks that the component exists in data and the UUID's match.

InfrastructureSystems.add_component! — Method.

add_component!(
               components::Components,
               component::T;
               kwargs...
              ) where T <: InfrastructureSystemsType

Add a component.

Throws ArgumentError if the component's name is already stored for its concrete type.

Throws InvalidRange if any of the component's field values are outside of defined valid range.

InfrastructureSystems.add_forecast! — Function.

add_forecast!(
              data::SystemData,
              ta::TimeSeries.TimeArray,
              component::InfrastructureSystemsType,
              label::AbstractString,
              scaling_factor::Union{String, Float64}=1.0,
             )

Add a forecast to a system from a TimeSeries.TimeArray.

See TimeseriesFileMetadata for description of scaling_factor.

InfrastructureSystems.add_forecast! — Function.

add_forecast!(
              data::SystemData,
              df::DataFrames.DataFrame,
              component::InfrastructureSystemsType,
              label::AbstractString,
              scaling_factor::Union{String, Float64}=1.0;
              timestamp=:timestamp,
             )

Add a forecast to a system from a DataFrames.DataFrame.

See TimeseriesFileMetadata for description of scaling_factor.

InfrastructureSystems.add_forecast! — Function.

add_forecast!(
              data::SystemData,
              filename::AbstractString,
              component::InfrastructureSystemsType,
              label::AbstractString,
              scaling_factor::Union{String, Float64}=1.0,
             )

Add a forecast from a CSV file.

See TimeseriesFileMetadata for description of scaling_factor.

InfrastructureSystems.add_forecast! — Method.

add_forecast!(data::SystemData, forecast)

Add a forecast.

Arguments

data::SystemData: infrastructure

forecast: Any object of subtype forecast

Throws ArgumentError if the forecast's component is not stored in the system.

InfrastructureSystems.add_forecasts! — Method.

add_forecasts!(
               ::Type{T},
               data::SystemData,
               metadata_file::AbstractString,
               label_mapping::Dict{Tuple{String, String}, String};
               resolution=nothing,
              ) where T <: InfrastructureSystemsType

Adds forecasts from a metadata file or metadata descriptors.

Arguments

::Type{T}: forecasted component type; may be abstract

data::SystemData: system

metadata_file::AbstractString: metadata file for timeseries that includes an array of TimeseriesFileMetadata instances or a vector.

resolution::DateTime.Period=nothing: skip forecast that don't match this resolution.

InfrastructureSystems.add_forecasts! — Method.

add_forecasts!(
               data::SystemData,
               timeseries_metadata::Vector{TimeseriesFileMetadata};
               resolution=nothing,
              )

Adds forecasts from a metadata file or metadata descriptors.

Arguments

data::SystemData: system

timeseries_metadata::Vector{TimeseriesFileMetadata}: metadata for timeseries

resolution::DateTime.Period=nothing: skip forecast that don't match this resolution.

InfrastructureSystems.check_contiguous_forecasts — Method.

Throws ArgumentError if the forecasts are not in consecutive order.

InfrastructureSystems.check_forecast_consistency — Method.

check_forecast_consistency(components::Components)

Throws DataFormatError if forecasts have inconsistent parameters.

InfrastructureSystems.clear_components! — Method.

Removes all components from the system.

InfrastructureSystems.compare_values — Method.

Recursively compares immutable struct values by performing == on each field in the struct. When performing == on values of immutable structs Julia will perform === on each field. This will return false if any field is mutable even if the contents are the same. So, comparison of any InfrastructureSystems type with an array will fail.

This is an unresolved Julia issue. Refer to https://github.com/JuliaLang/julia/issues/4648.

An option is to overload == for all subtypes of PowerSystemType. That may not be appropriate in all cases. Until the Julia developers decide on a solution, this function is provided for convenience for specific comparisons.

InfrastructureSystems.convert_type — Method.

Converts an object deserialized from JSON into a Julia type, such as NamedTuple, to an instance of T. Similar to Base.convert, but not a viable replacement.

InfrastructureSystems.convert_type — Method.

Creates a Forecast object by decoding the data that was in JSON. This data stores the values for the field contributingdevices as UUIDs, so this will lookup each device in devices.

InfrastructureSystems.deserialize_components — Method.

Deserializes components defined in InfrastructureSystems. Parent modules should override this by changing the component type and module.

InfrastructureSystems.from — Method.

from(forecast::Forecast, timestamp)

Return a forecast truncated starting with timestamp.

InfrastructureSystems.from_file — Method.

Constructs Hdf5TimeSeriesStorage from an existing file.

InfrastructureSystems.from_json — Method.

Deserializes a InfrastructureSystemsType from a JSON filename.

InfrastructureSystems.from_json — Method.

Deserializes a InfrastructureSystemsType from String or IO.

InfrastructureSystems.generate_initial_times — Method.

generate_initial_times(
                       component::InfrastructureSystemsType,
                       interval::Dates.Period,
                       horizon::Int,
                      )

Generates all possible initial times for the stored forecasts. This should return the same result regardless of whether the forecasts have been stored as one contiguous array or chunks of contiguous arrays, such as one 365-day forecast vs 365 one-day forecasts.

Throws ArgumentError if there are no forecasts stored, interval is not a multiple of the system's forecast resolution, or if the stored forecasts have overlapping timestamps.

InfrastructureSystems.generate_initial_times — Method.

generate_initial_times(data::SystemData, interval::Dates.Period, horizon::Int)

Generates all possible initial times for the stored forecasts. This should return the same result regardless of whether the forecasts have been stored as one contiguous array or chunks of contiguous arrays, such as one 365-day forecast vs 365 one-day forecasts.

Throws ArgumentError if there are no forecasts stored, interval is not a multiple of the system's forecast resolution, or if the stored forecasts have overlapping timestamps.

InfrastructureSystems.get_abstract_subtypes — Method.

Returns an array of abstract types that are direct subtypes of T.

InfrastructureSystems.get_all_concrete_subtypes — Method.

Returns an array of all concrete subtypes of T.

InfrastructureSystems.get_component — Method.

get_component(
              ::Type{T},
              components::Components,
              name::AbstractString
             )::Union{T, Nothing} where T <: InfrastructureSystemsType

Get the component of concrete type T with name. Returns nothing if no component matches.

See get_components_by_name if the concrete type is unknown.

Throws ArgumentError if T is not a concrete type.

InfrastructureSystems.get_component_types_raw — Method.

Return an iterable of component types deserialized from JSON.

InfrastructureSystems.get_components — Method.

get_components(
               ::Type{T},
               components::Components,
              )::FlattenIteratorWrapper{T} where T <: InfrastructureSystemsType

Returns an iterator of components. T can be concrete or abstract. Call collect on the result if an array is desired.

See also: iterate_components

InfrastructureSystems.get_components_by_name — Method.

get_components_by_name(
                       ::Type{T},
                       components::Components,
                       name::AbstractString
                      )::Vector{T} where T <: InfrastructureSystemsType

Get the components of abstract type T with name. Note that InfrastructureSystems enforces unique names on each concrete type but not across concrete types.

See get_component if the concrete type is known.

Throws ArgumentError if T is not an abstract type.

InfrastructureSystems.get_components_raw — Method.

Return an iterable of components as NamedTuples deserialized from JSON.

InfrastructureSystems.get_concrete_subtypes — Method.

Returns an array of concrete types that are direct subtypes of T.

InfrastructureSystems.get_forecast — Method.

get_forecast(
             ::Type{T},
             component::InfrastructureSystemsType,
             initial_time::Dates.DateTime,
             label::AbstractString,
             horizon::Int,
            ) where T <: Forecast

Return a forecast for a subset of the time series range stored for these parameters.

InfrastructureSystems.get_forecast — Method.

get_forecast(
             ::Type{T},
             component::Component,
             initial_time::Dates.DateTime,
             label::AbstractString,
            ) where T <: Forecast

Return a forecast for the entire time series range stored for these parameters.

InfrastructureSystems.get_forecast_counts — Method.

get_forecast_counts(data::SystemData)

Return a tuple of counts of components with forecasts and total forecasts.

InfrastructureSystems.get_forecast_values — Method.

get_forecast_values(
                    ::Type{T},
                    component::InfrastructureSystemsType,
                    initial_time::Dates.DateTime,
                    label::AbstractString,
                   ) where T <: Forecast

Return a TimeSeries.TimeArray where the forecast data has been multiplied by the forecasted component field.

InfrastructureSystems.get_forecasts_horizon — Method.

Return the horizon for all forecasts.

InfrastructureSystems.get_forecasts_interval — Method.

Return the time delta between the first two stored forecasts. if less than two are stored, return Dates.Second(0).

InfrastructureSystems.get_forecasts_resolution — Method.

Return the resolution for all forecasts.

InfrastructureSystems.get_period_columns — Method.

Return the column names that specify the Period.

InfrastructureSystems.get_step_time — Method.

Return a DateTime for the step between values as specified by the period in the file.

InfrastructureSystems.get_timeseries_format — Method.

Return the timeseries format used in the CSV file.

InfrastructureSystems.get_timestamp — Method.

Return a Dates.DateTime for the row in the CSV file.

InfrastructureSystems.get_unique_timestamps — Method.

Return a vector of dicts of unique timestamps and their counts.

InfrastructureSystems.get_uuid — Method.

Gets the UUID for any PowerSystemType.

InfrastructureSystems.get_value_columns — Method.

Return the column names with values.

InfrastructureSystems.get_value_columns — Method.

Return the column names with values (components).

InfrastructureSystems.head — Method.

head(forecast::Forecast)
head(forecast::Forecast, num)

Return a forecast with only the first num values.

InfrastructureSystems.increment_count — Method.

Increments the count of a log event.

InfrastructureSystems.iterate_components — Method.

iterate_components(obj) where T <: InfrastructureSystemsType

Iterates over all components.

Examples

for component in iterate_components(obj)
    @show component
end

See also: get_components

InfrastructureSystems.make_forecast! — Method.

make_forecast!(timeseries_metadata::TimeseriesFileMetadata;
               resolution=nothing)

Return a vector of forecasts from TimeseriesFileMetadata.

Arguments

timeseries_metadata::TimeseriesFileMetadata: metadata

resolution::{Nothing, Dates.Period}: skip any forecasts that don't match this resolution

InfrastructureSystems.prepare_for_removal! — Method.

This function must be called when a component is removed from a system.

InfrastructureSystems.prepare_for_serialization! — Method.

prepare_for_serialization!(data::SystemData, filename::AbstractString)

Parent object should call this prior to serialization so that SystemData can store the appropriate path information for the time series data.

InfrastructureSystems.read_time_series — Function.

read_time_series(file_path::AbstractString, component_name=nothing)

Return a TimeArray from a CSV file.

Pass component_name when the file does not have the component name in a column header.

InfrastructureSystems.read_time_series — Method.

This version of the function supports the format where there is no column header for a component, so the component_name must be passed in.

InfrastructureSystems.read_time_series — Method.

This version of the function only has component_name to match the interface. It is unused.

Set start_datetime as a keyword argument for the starting timestamp, otherwise the current day is used.

InfrastructureSystems.read_time_series — Method.

Return a TimeSeries.TimeArray representing the CSV file.

This version of the function only has component_name to match the interface. It is unused.

InfrastructureSystems.read_time_series_metadata — Method.

Reads forecast metadata and fixes relative paths to the data files.

InfrastructureSystems.remove_component! — Method.

remove_component!(
                  components::Components,
                  component::T,
                 ) where T <: InfrastructureSystemsType

Remove a component by its value.

Throws ArgumentError if the component is not stored.

InfrastructureSystems.remove_component! — Method.

remove_component!(
                  ::Type{T},
                  components::Components,
                  name::AbstractString,
                  ) where T <: InfrastructureSystemsType

Remove a component by its name.

Throws ArgumentError if the component is not stored.

InfrastructureSystems.remove_components! — Method.

remove_components!(
                   ::Type{T},
                   components::Components,
                  ) where T <: InfrastructureSystemsType

Remove all components of type T.

Throws ArgumentError if the type is not stored.

InfrastructureSystems.remove_forecast! — Method.

remove_forecast!(
                 ::Type{T},
                 data::SystemData,
                 component::InfrastructureSystemsType,
                 initial_time::Dates.DateTime,
                 label::String,
                ) where T <: Forecast

Remove the time series data for a component.

InfrastructureSystems.remove_forecast_internal! — Method.

remove_forecast_internal!(
                          ::Type{T},
                          component::InfrastructureSystemsType,
                          initial_time::Dates.DateTime,
                          label::AbstractString,
                         ) where T <: ForecastInternal

Removes the metadata for a forecast. The caller must also remove the actual time series data.

InfrastructureSystems.replace_iterator — Method.

Replace the iterator, maintaining the cached dict.

InfrastructureSystems.reset_iterator — Method.

Reset the iterator for cases where underlying arrays have changed.

InfrastructureSystems.separate_type_and_parameter_types — Method.

Return a Tuple of type and parameter types for cases where a parametric type has been encoded as a string. If the type is not parameterized then just return the type.

InfrastructureSystems.set_component! — Method.

set_component!(
               metadata::TimeseriesFileMetadata,
               components::Components,
               mod::Module,
              )

Set the component value in metadata by looking up the category in module. This requires that category be a string version of a component's abstract type. Modules can override for custom behavior.

InfrastructureSystems.strip_module_name — Method.

Strips the module name off of a type.

InfrastructureSystems.supertypes — Method.

Returns an array of all super types of T.

InfrastructureSystems.tail — Method.

tail(forecast::Forecast)
tail(forecast::Forecast, num)

Return a forecast with only the ending num values.

InfrastructureSystems.to — Method.

to(forecast::Forecast, timestamp)

Return a forecast truncated after timestamp.

InfrastructureSystems.to_json — Method.

JSON Serializes a InfrastructureSystemsType to an IO stream in JSON.

InfrastructureSystems.to_json — Method.

Serializes a InfrastructureSystemsType to a JSON file.

InfrastructureSystems.to_json — Method.

Serializes a InfrastructureSystemsType to a JSON string.

InfrastructureSystems.type_to_symbol — Method.

Converts a DataType to a Symbol, stripping off the module name(s).

InfrastructureSystems.validate_components — Method.

validate_components(components::Components)

Iterates over all components and throws InvalidRange if any of the component's field values are outside of defined valid range.

InfrastructureSystems.validate_struct — Method.

Validates a struct.

InfrastructureSystems.when — Method.

when(forecast::Forecast, period::Function, t::Integer)

Refer to TimeSeries.when(). Underlying data is copied.

JSON2.read — Method.

Enables JSON deserialization of TimeSeries.TimeArray. The default implementation fails because the data field is defined as an AbstractArray. Deserialization can't determine the actual concrete type.

JSON2.write — Method.

Enables JSON deserialization of Dates.Period. The default implementation fails because the field is defined as abstract. Encode the type when serializing so that the correct value can be deserialized.